Factor out FilterSelect and WeightsConfigSection components Convert from aphrodite to material-ui style creation No functional changes
